Refactor user table variable user_id and empNum convert long into int

+++ backend/daoumarket/src/main/java/com/daoumarket/dao/FavoriteDao.java
@@ -27,7 +27,7 @@ public class FavoriteDao implements IFavoriteDao {
 	}
 
 	@Override
-	public List<Favorite> getFavoriteList(long userId) {
+	public List<Favorite> getFavoriteList(int userId) {
 		return sqlSession.selectList(ns + "getFavoriteList", userId);
 	}
