+++ backend/daoumarket/src/main/java/com/daoumarket/service/ISearchService.java
@@ -8,5 +8,5 @@ import com.daoumarket.dto.SearchInsertResponse;
 public interface ISearchService {
 	public ResponseEntity<BasicResponse> insertSearch(SearchInsertResponse searchInsertResponse);
 	public ResponseEntity<BasicResponse> deleteSearchHistory(long searchId);
-	public ResponseEntity<BasicResponse> getSearchHistory(long userId);
+	public ResponseEntity<BasicResponse> getSearchHistory(int userId);
 }
